requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
supabase>=2.6.0
python-dotenv>=1.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
groq>=0.8.0
cloudinary>=1.40.0
python-socketio>=5.11.2
python-multipart>=0.0.9
pydantic>=2.7.4
pydantic-settings>=2.3.4

